Fix un-necessary VMSS instance rolling request (#12590)
Rolling VMSS instances when upgrade mode is set to `Automatic` is only
allowed of the image version is set to `latest`, otherwise we get the
following error:

```
The OS Rolling Upgrade API cannot be used on a Virtual Machine Scale Set unless the Virtual Machine Scale Set has some unprotected instances which have imageReference.version set to latest.
```

This change makes sure we only roll instances if the image version is
`latest`. The VMSS will roll instances on its own for non-latest images
as long as upgrade_type is set to `Automatic`.