new resource "azurerm_data_factory_custom_dataset" (#12484)
generic resource about dataset

notes:

according to the service team, structure is deprecated and schema is prefered. So in this resource, I didn't imeplement structure property.
schema is type interface, according to the service team, it could be either a object or an array. So I also made it a json string. In the acctest, both cases are covered.