Refine the ExpandedConfig field type
Currently, the sdk encoder/decoder only supports non-pointer type. The
`ExpandedConfig` is to be used as the "typed" schema which is decoded
from the TF schema, and then later used to cast to/from the SDK type.

This commit modify the field type of `ExpandedConfig` to be non-pointer,
and made other changes accordingly. Additionally, it adds the `tfschema`
tag for the fields.