"azurerm_data_factory_pipeline" - supports "concurrency", "elapsed_time_metric_duration" and "run_dimension" (#12685)
fix #11506

"run_dimension" is a key value pair, but the value could be a string or an expression type, so made it TypeList. It's used by ChainingTrigger.

ChainingTrigger allows the referenced pipeline to depend on other pipeline runs based on runDimension Name/Value pairs. Upstream pipelines should declare the same runDimension Name and their runs should have the values for those runDimensions. The referenced pipeline run would be triggered if the values for the runDimension match for all upstream pipeline runs.