Handle nil values for AllowBlobPublicAccess (#12689)
There is a certain edge case that could result in the Azure API returning a null value for AllowBLobPublicAccess.
Since the field is a pointer, this gets marshalled to a nil value instead of a boolean. 

This change causes the provider to set this to the default value, but since this is a breaking change we'll enforce this in 3.0.