Backwards compatible fix for Cluster Variable
https://github.com/hashicorp/terraform-provider-azurerm/issues/13099

This default value causes clusters to be rebuilt on existing
deployments. Simply removing the default value should be sufficient for
preventing the rebuilds from occuring.

If someone needs to explicitly set this to false for some reason, they
can still do that manually. This is a preferable situation to existing
users not being able to upgrade without a complete cluster rebuild.