azurerm_logic_app_workflow - support for enabled & access_control (#13265)
This PR is to add support for new properties state and accessControl for azurerm_logic_app_workflow.

API failed to create Logic App Workflow with "accessControl.triggers". Per the error message, it complains "OpenAuthenticationAccessPolicy.type" should be "AAD" while creating this resource with "accessControl.triggers", which means "OpenAuthenticationAccessPolicy.type" should be writeable. But after checked swagger, I found this property is marked as read-only. Service team confirmed it's a bug in swagger. So I filed an issue on rest api spec and didn't involve it in this PR. Once the bug is fixed, I will submit another PR to add it.

--- PASS: TestAccLogicAppWorkflow_empty (261.74s)
--- PASS: TestAccLogicAppWorkflow_requiresImport (279.18s)
--- PASS: TestAccLogicAppWorkflow_tags (465.55s)
--- PASS: TestAccLogicAppWorkflow_integrationAccount (590.05s)
--- PASS: TestAccLogicAppWorkflow_parameters (240.56s)
--- PASS: TestAccLogicAppWorkflow_accessControl (374.61s)