azurerm_mssql_database: Attempt to work around SKU change constraints for replicating databases
When databases are replicating, the primary cannot have a SKU belonging
to a higher service tier than any of its partner databases. To work
around this, we'll try to identify any partner databases that are
secondary to this database, and where the new SKU tier for this database
is going to be higher, first upgrade those databases to the same
`sku_name` as we'll be changing this database to. If that sku is
different to the one configured for any of the partner databases, that
discrepancy will have to be corrected by the resource for that database.

That might happen as part of the same apply, if a change was already
planned for it, else it will only be picked up in a second plan/apply.

For the best experience, configurations should use the same SKU for a
primary database and any secondary databases that are replicating from
it.