storage_data_lake_gen2_path note about ACL inheritance
Azure will automatically apply the default ACL to all nested directories
(which is expected) but terraform knows nothing about it and tries to
revert to what it believes to be right.

At this point it's not possible for terraform to handle this isse since
we cannot tell if the ACL change took place because the of inheritance
or because someone else changed the permissions manually.