Throw error if there is a missing capability when enabling mongo
When we enable capability `MongoDBv3.4` azure also sets `EnableMongo`.
As a result next time terraform runs it sees both capabilities enabled,
and since only `MongoDBv3.4` is enabled in the config it will try to
change the capabiltiies to match what it has in the config.  In this
case it means destroying the database and building it again, since the
`capabilities` setting has `ForceNew` set to true.

This commit uses a CustomizeDiff function to enforce this during plan.