`azurerm_lb`, `azurerm_lb_backend_address_pool`, `azurerm_network_interface` - add supports for gateway LB (#13559)
This PR adds support for the Gateway Loadbalancer. Checkout the Swagger definition for examples.

There are two principals in the use case of the gateway LB: the provider and the consumer. The provider is the one that provisions the gateway loadbalancer, where the loadbalancer's sku is set to Gateway and its BAP has the tunnel_interface defined. The consumer is the one that directs its traffic to the provider gateway LB. There are currently two resources can consume (point to) the gateway LB: the azurerm_lb and the azurerm_network_interface. For these two resources, this PR has added the gateway_load_balancer_frontend_ip_configuration_id to their forntend configuration block (arguably, the name of the property is a bit too verbose, while I didn't figure out another better name 😅 )

Probably related to #13453