`azurerm_firewall_policy - add supports for the `insights` (#14004)
This PR add supports of the insights for the azurerm_firewall_policy.

This feature logs additional information in the backend and these workspaces are used for logging. Firewall policy can be associated with one or more Firewalls. The cost of logging data across regions is expensive. Imagine a firewall running in India West logging to East US. It’s way too expensive. This feature provide an ability to control where the Firewall will log based on the region where it’s deployed. If Firewall is in an unspecified region, the default workspace ID kicks in.