key_vault_key - add the cureve, x, y, public_key_pem, public_key_openssh (#13934)
This attempts to close match the attributes that the terraform tls_private_key provides. Specifically the OpenSSH and PEM format for the public keys.

At the moment the only access to the public keys are in the form of key type, key curve and then base 64 encoded N/E/X/Y values. However these aren't too helpful at runtime for terraform as you can't meaningfully manipulate these in a way that may be necessary.

For my particular case I need the OpenSSH format public key so that I can use the key vault backed key with SSH authentication. The OpenSSH format key is added to a VM as an authorized_key and the key vault key is then used for SSH authentication (yes this works and is implemented).

At the moment there's no built-in support for P256 Curves, so those keys will not be formatted but should silently fail.