removed hostpool_reference from acceptance test
The hostpool_reference fails on permissions during the acceptance test. The same configuration works perfectly using a dev build of the provider outside of the acceptance test so must be something specific to the acceptance test execution that is causing an issue. Removing as the issue is an acceptance test issue rather than a resource issue.