`r\purview_account`: Add support for `managed_resource_group_name` (#14217)
When creating a Purview Account, a new resource group will be created automatically together with it. The name of this resource group is decided by server by default, and user can override it. This change adds support for the override capability.
The update API accepts the managedResourceGroupName, but it doesn't change the actual value, so the property is set to ForceNew
Validation is azure.ValidateResourceGroupName since it needs to be a valid resource group name