New attribute for `r/virtual_desktop_application_group` (#14227)
Co-authored-by: Andy.McCullough <andy.mccullough@ultima.com>
Co-authored-by: kt <kt@katbyte.me>

This pull adds a new default_desktop_display_name attribute for the virtual_desktop_application_group resource

When creating a new Desktop application group - a default desktop is created named sessionDesktop with a display name of sessionDesktop. There is no means through the application group API to modify this.

This leverages the Desktops API to allow users to configure the display name of the default desktop resource in-line as part of the application group resource when a Desktop type is created.

I did consider a new resource to implement this, however the DesktopsClient is limited in that it only supports the updating of Desktops not the Creation or Deletion, so didn't seem appropriate.