Upgrade Api Management from 2020-12-01 to 2021-08-01 (#14312)
Microsoft added a large amount of functionality in 2021-01-01-preview for API Management. This PR is to move us to the latest stable API release, which will allow for adding functionality to resolved issues such as #12021 and #13445. It updates all the constants which were renamed.

The only potential regression is that Sample is no longer a property in the OperationRepresentation schema. It has been replaced with a map of Examples. I can update the schema, but I'm not sure how to deal with the value property of the ParameterExampleContract. If someone can help me with this, I'll push the code for the Examples.

My plan is to uplift the features in separate PRs.