r/consumption_budget_subscription: adding a v2 state migration
Whilst #14687 attempted to fix the issue in the original state migration unfortunately
doing so would lead to multiple possible values for this field (the "fixed" value for
users who skipped releases containing this bug, and a "broken" value for users who didn't).

As such this commit introduces a new state migration from v1 -> v2 which parses the
Subscription ID from the Resource ID, meaning that value will be the same (and correct) in
both cases.