Add managed_resources block to purview_account 14814 (#14865)
Co-authored-by: Tom Harvey <tombuildsstuff@users.noreply.github.com>
Co-authored-by: Noel Jones <noel-jones@officedepot.eu>

Supports the managed_resource block requested in my enhancement resquest issue #14814

This is my first Terraform change therefore grateful for consideration in respect of any missed steps but hopefully I have covered it all!

With respect to the attribute names, I have kept them in sync with the Rest API - I considered it better to be in sync than to add "_id" to the attributes, however, I will do this if this is considered best practice for resource IDs. The documentation states that they are IDs.

Tested within our infrastructure, plan results in the following output against an existing resource and equivalent for a new resource.