`azurerm_network_watcher_flow_log`: the `name` property can now be configured for new resources. (#15016)
This PR expose the name for the azurerm_network_watcher_flow_log, as an enhancement after #12533. It is implemented in a backward compatible manner.
I've tested locally that I can import a flow log resource that is created without setting the name via either the old import format and the new import format, both has no plan diff.