azurerm_monitor_action_group: fix the format issue of the workspace_id property for itsm_receiver (#15422)
Fix issue #9405. Currently, the workspace_id property in the terraform provider is required to be a UUID. In fact, the format required by the API is "{subscriptionid}|{workspaceid}", which has been confirmed by the service team. This is an example of workspace_id . Since the monitor action group API returns success even though the workspace_id is set to a single UUID, I filed an API issue to track it. And update the "ValidateFunc" of workspace_id to allow the user to set the correct workspace_id value in terraform provider.