Set key vault id on key import issue #12357 (#15670)
Co-authored-by: Noel Jones <noel-jones@officedepot.eu>

Use the nested importer i the same manner as for certificates and secrets, which sets the key vault id.

Tested import of key created via console and import of key previously created through Terraform. Both imported successfully and Terraform did not attempt to recreate (though it did 'update' the console created key because of the different order of the key_opts but that is another matter common to many resource sets).

I don't believe any further documentation or test files need updating.