azurerm_mssql_database_vulnerability_assessment_rule_baseline: fix the resource is being replaced every new tf apply issue (#14759)
The purpose of this PR:

Fix Issue #14695. The issue is that the resource azurerm_mssql_database_vulnerability_assessment_rule_baseline is always created when the "terraform apply" runs every time, even if the tf config is not changed. The server_vulnerability_assessment_id property is ForceNew in terraform. So, when the value of the property is changed, the resource would be re-created. In this case, the root cause is that the DatabaseVulnerabilityAssessments API returns the VulnerabilityAssessmentName of id in uppercase, which changes the property value. So, I submitted this PR as a workaround for the issue. And I created an issue for Azure API to fix it completely. And this PR should be reverted after the API issue is fixed.