Correct documentation of boundary conditions for azurerm_storage_share.quota property
The previous phrasing 'less than' implies to me that the upper bound is exclusive. From experience with my own resources `5120` is an acceptable value for standard storage accounts and checking the validation code in the resource the `102400` limit also appears to be inclusive. Therefore I've tweaked the phrasing to correclty indicate that the boundary conditions are inclusive.