`azurerm_dns_x_record` - Enlarge the read timeout from 5min to 10min
DNS has a throttle limit of 500 read (GET) operations per 5 minutes. DNS record is a resource that users might provision in a large scale, which likely triggers this throttle. When throttle is hit, the 5min read timeout is not long enough for the retry. So extending this to 10min to at least allow one time retry (probably allow more than 500 but less than 1000 scale).