remove validation of app service sticky setting names
This validation is too strict. The Azure Portal allows the following
name for both app settings and connection strings:

    A test: !@#$%^&*()_+-=' ";/?

I can't find a reference for what the valid character set would be,
and the names in the app_settings attribute aren't validated at all,
so we may as well remove this validation.

Fixes #16612