sdk: extracting the Plugin SDKv2 schema types to their own interface
This provides the foundation for switching these out on a per-resource basis later - likely to
an intermediate type which compiles down to Plugin SDKv2 and/or Plugin Framework, rather than
Plugin Framework types directly - given there's common behaviours we want to expose.

For now defining this as a separate interface allows auto-completion to work for the resources
whilst also allowing us (later on) to gradually roll out the replacement schema types.