r/managed_disk: support for Expand Without Downtime
This extends the work done by @kazimierzbudzyk to ensure we check the Virtual Machine and Managed Disk to confirm
that they support Expand Without Downtime (and fallback to shutting the machine down if not).

Unfortunately we're unable to output a warning either way at this time (since that'd require Framework) - as such
this PR includes adding notes to the documentation calling out we'll best-effort this.