Default to 1 WoY in long term retention policy
`week_of_year` is only necessary if `yearly_retention` is enabled.
Valid values of `week_of_year` go from 1 to 52.

When `yearly_retention` is not set, `PT0S`, the default value returned
by the API for `week_of_year` is 0. However, this is not a valid value.
Therefore, in this case, when flattening the response, the default
`week_of_year` value is set to 1.

Fixes #13035