`external_id` needs to reference tenant id when adding AAD group
In the documentation it gives an example in the description saying `aad://<tenant>.onmicrosoft.com/groups/<group object id>`. For the group to actually work in APIM `external_id` needs to be `aad://<tenant id>/groups/<group object id>`. If not, I've experienced group membership not picked up by APIM.