`azurerm_key_vault_{key|secret|certificate}` - Directly using `key_vault_id` during read and delete
The `key_vault_id` is a required, force new property, which is always a constant in the state file if the resource is properly created/imported via Terraform. Especially, for import, this property is constructed by the importer and set to the state prior to calling the `Read` for all the affected 3 resources. Therefore, there is actually only one place that is required to transmute the key vault URI to its resource id: the importer. While for the others:

- Create/Update: it is always available in the plan input
- Read/Delete: it is always available in the state

This PR removes the transmuation code in the read and delete, but directly use the `key_vault_id` from state.