Bugfix: fall back to retrieving objectID of authenticated principal from MS Graph
This is still needed when authenticating between tenants, i.e. the
authentication principal (application or user) is homed in a different
tenant to the authorizing tenant, as the `oid` claim is often missing.

Additionally, fall back to configuring tenantID and clientID from
user-supplied values in the event those claims are also absent from the
access token.