d/client_config: removing `environment_name` for the moment
Unfortunately the name of the environment isn't guaranteed not to change, and has done so
several times in the past (Public, AzurePublicCloud -> AzureCloud) - as such this
isn't something we can provide a compatibility guarantee for at this time.

It's worth noting that we source this information from the MetaData service, so it's possible
for this to change outside of Terraform too, should the Metadata endpoint rename this.

Instead we'll want to add a Data Source to surface the relevant information /for/ the current
environment, for example the Private Link name, but that's something we'll need to do once
the change to this Data Source is reverted (since this isn't something we can provide a compatibility
guarantee for at this time, per above).