`nfsv3_enabled` note update (#21654)
Updated the note for the requirements of the `nfsv3_enabled` argument in the azure storage account docs. Choosing `GRS`, `ZRS`, `GZRS` or `RAGZRS` for the `account_replication_type` argument causes an error while applying changes. The error takes one of the following forms depending on the value chosen:
│ Error: creating Azure Storage Account ... : Failure sending request: StatusCode=400 -- Original Error: Code="InvalidRequestPropertyValue" Message="The value 'True' is not allowed for property isNfsv3Enabled."
│ Error: reading share properties for AzureRM Storage Account ... : Failure responding to request: StatusCode=404 -- Original Error: autorest/azure: Service returned an error. Status=404 Code="ResourceNotFound" Message="The specified resource does not exist.\n ... "
Only `LRS` and `RAGRS` are valid values.