Improve error messages for resource IDs with non-canonical casing
The Azure APIs and portal are not entirely consistent with casing in resource IDs. This is noted in ParseAzureResourceID where `resourcegroups` are special-cased to work around inconsistency. This behaviour can cause confusion when importing resources, as the error message `ID was missing the 'ruleSet' element` (for example) appears non-sensical if the user doesn't notice the capitalisation difference. This is exacerbated by the `az` cli returning the ID as provided when provided to show subcommands, meaning you can copy/paste an ID from the command-line output (or certain places in the portal) and be told it's missing a component.

This change does not alter the parsing behaviour, the case sensitivity of path segments is still maintained. It does, however, add an additional check to the error case of popping a segment from the path. If the path does not have a segment of that name, this change will iterate over the path segments and validate if any are a case-insensitive match. If so, this provides a more diagnostic error message.

I believe this is an appropriate efficiency trade-off of computer versus human time, the error case will be hit relatively infrequently, and resource IDs are generally a small number of segments. The impact on increased computation needed should be outweighed by allowing users to navigate this confusing error case more easily.