Add option to re-use host_name with `dns_zone with soa_record` creation
When a DNS zone is created in Azure, there is a default SOA record which is
created.

In this provider, a dns_zone with soa_record is done by:
1. creating the dns_zone
2. overwriting the soa_record of the zone with the data supplied in
   the soa_record config block

The `soa_record` block makes it mandatory to add the host_name argument
when it is specified. This is NOT desired in most cases since the host_name
is the MNAME SOA field which is the name of the primary nameserver and is
set by Azure.

This commit attempts to fix this issue by:
1. allowing the `host_name` to be omitted when `soa_record` is specified
2. the host_name value will be derived from the original SOA record when
   the zone is created.

This issue is partly a bug in Azure API since it is is mentioned in the
[documentation](https://learn.microsoft.com/en-us/azure/dns/dns-zones-records#soa-records)
that it should not be possible to update the 'host' property of the SOA record.