azurerm_synapse_firewall_rule: fix rule name regex
The rule name regex is actually more lenient. To quote azure:
```
The firewall rule name cannot be empty, it cannot contain a
slash (/) or a backslash (\), can't end with '.', and can't
contain '<,>,*,%,&,:,\,/,?'. Additionally, the firewall rule
name cannot exceed 128 characters.
```

As can be seen on the validation image:
![image](https://github.com/hashicorp/terraform-provider-azurerm/assets/17879459/eefd1a6e-5dd0-407a-89e0-bbd97b8459d6)

This fixes that

Signed-off-by: Frank Villaro-Dixon <frank.villarodixon@merkle.com>