Allow AzureRM to skip configuration/authentication
By default, the AzureRM Provider will attempt to authenticate with Azure on first use.
In some rare cases, this is not desired. Example, a module may contain resources/data sources
from multiple cloud provider but the module is being configured to creat/use only non-Azure
resources/data sources. In this case, the module requires the AzureRM provider to be present
but Terraform will not effectively use it.

In this pull request, we introduce a new: `skip_provider_configuration` configuration
parameter (or use the ARM_SKIP_PROVIDER_CONFIGURATION environment variable) to support
this use case.