New Tool: `update-go-azure-sdk`
This tool is intended to update the `hashicorp/go-azure-sdk` dependency within the Provider - and then update the codebase
to take advantage of any new API Versions that are present within this version of `hashicorp/go-azure-sdk`, if used in the
Provider and the API versions are compatible (e.g. the tests pass when updating the API Version).

This tool makes a considerable number of assumptions about the working directory and environment this is being run in and
is primarily intended to be run in automation - and whilst will not be automatically configured initially, in time should
mean that both SDK upgrades and new API Version upgrades are (mostly) automated.

In the event that an SDK/API Version is _not_ compatible, the automated process will fail - and manual intervention is
required - which is desired (since this should be the edge-case and/or requires additional thought).