feat: Enable Workload Profile for Container App Environment
This commit enables the option to enable Workload Profile for a Container App Environment. The `workload_profiles_enabled` field has been added to the `ContainerAppEnvironmentModel` struct and the `Arguments()` function. Additionally, in the `Create()` function, if `WorkloadProfileEnabled` is set to true, a new workload profile named "Consumption" will be created. In the `Read()` function, the value of `WorkloadProfileEnabled` will be determined based on whether there are any workload profiles associated with the environment.