Update vendors with latest go-sdk
It looks like the SDK was updated on Jan 18th 2024 (3 days ago),
and Openshift changes with vendor landed on Jan 19th. It was very
likely the case that the openshift vendor changes were not
regenerated when the SDK changed and the new resource was merged
here https://github.com/hashicorp/terraform-provider-azurerm/pull/24375.

This commit updates the vendors directory with latest changes.