r/resource_group: updating the Exists and Delete checks to use `hashicorp/go-azure-sdk`
This commit updates the Exists and Delete check functions to use `hashicorp/go-azure-sdk`
rather than `Azure/azure-sdk-for-go` to enable cross-subscription test assertions to be used.

Due to the `azurerm_resource_group` resource still using the Resource ID from the Azure
API for historical reasons, this commit intentionally parses the Resource ID insensitively
during the Exists and Delete functions - but this is a temporary solution until the Resource
is updated in the future.