doc: Update description for resource_group_name argument
The way the resource is written currently, the `resource_group_name` argument is specifically needing the Resource Group of the target Recovery Services Vault. To be honest, I'm not sure why the `resource_group_name` and `recovery_vault_name` arguments are being used instead of the more common/modern approach of using a `recovery_services_vault_id` or `recovery_vault_id` argument.