Add Boolean option for `EnablePrivateLinkFastPath`
This was supposed to be introduced in
https://github.com/hashicorp/terraform-provider-azurerm/pull/20619,
however it wasn't and `ExpressRouteGatewayBypass` was introduced. This
is however also a required option for `EnablePrivateLinkFastPath` to
work.

This is a very simple Bool, that is referenced in the API specs, since https://learn.microsoft.com/en-us/azure/templates/microsoft.network/2022-07-01/expressroutegateways?pivots=deployment-language-terraform