Add `EnablePrivateLinkFastPath` to the Express Route Connection resource (#25596)
* Add Boolean option for `EnablePrivateLinkFastPath`

This was supposed to be introduced in
https://github.com/hashicorp/terraform-provider-azurerm/pull/20619,
however it wasn't and `ExpressRouteGatewayBypass` was introduced. This
is however also a required option for `EnablePrivateLinkFastPath` to
work.

This is a very simple Bool, that is referenced in the API specs, since https://learn.microsoft.com/en-us/azure/templates/microsoft.network/2022-07-01/expressroutegateways?pivots=deployment-language-terraform

* add `enable_private_link_fast_path` to tests

* Add documentation for `enable_private_link_fast_path`

I copied the text from https://learn.microsoft.com/en-us/azure/templates/microsoft.network/expressroutegateways?pivots=deployment-language-terraform#expressrouteconnectionproperties-2

* Rename `enable_private_link_fast_path` to `private_link_fast_path_enabled`

* Add option to complete tests

* Refactor by suggestion

* use `pointer` helper

* update documentation

* Guard `private_link_fast_path_enabled` being set

The documentation states, that `express_route_gateway_bypass_enabled`
needs to be enabled for `private_link_fast_path_enabled` to work. Thus
we're guarding the `Create` and `Update` function to error, in case the
dependency isn't met.

* Add `PrivateLinkFastPath` to Update function

* Update attribute name to match Specs

* Add dependency to test `express_route_gateway_bypass_enabled`