appservice: conditionally rewriting the value for `virtual_network_subnet_id`
This fixes https://github.com/hashicorp/terraform-provider-azurerm/issues/25832 by ensuring that the Subnet ID is recased
meaning that it'll be in the correct format.

From our side since we're validating the user input for this field, it's hard for us to write a test-case for this one since
it'd be caught by the validation - but this is caused by user input for this field (provisioned through an ARM Template etc)
providing this value without the prefix '/', which the API should be requiring (but isn't).