azurerm_nginx_deployment - add ForceNew to nested resources (#26298)
A bug was found when trying to update a deployment's private IP. The expected behavior is that the deployment
should be recreated. However, the actual behavior was that terraform plan showed an in-place update.

This is odd because the frontend_private field, the parent field for the private IP address has ForceNew set
to true. However, there seems to be a bug where if the field is a list type and the Elem is of type
pluginsdk.Resource, the ForceNew does not get propogated down to the child fields.

This change adds ForceNew: true to the nested fields whose parents have ForceNew: true.