`azure_nginx_deployment`: Use the `standardv2_Monthly` SKU (#27395)
* Document new SKU on NGINXaaS

* Change NGINXaaS SKU to standardv2_Monthly

NGINXaaS now supports a newer SKU
`standardv2_Monthly` and the `standard_Monthly`
SKU has been deprecated.

This change updates the acceptance tests to use
the new SKU.

* Drop possible SKU value fron NGINXaaS datasource

The data source documentation does not need to
tell the user for the possible values of a field
as it is not an input and the user will see the
value once they run the datasource itself.

* Update docs with new NGINXaaS SKU

With the deprecation of standard_Monthly SKU, we
want users to start using the new
`standardv2_Monthly` SKU while we work on retiring
the old one. With that in mind, we'd like our
users to see just latest supported SKUs.