`azurerm_nginx_configuration`: Update config handlers to deprecate content of protected_files
The service provider has never returned the
contents of a protected file in the API response.
The latest swagger splits up the request,response
objects on protected files, and this commit
incorporates the swagger change while deprecating
the `content` field on the protect file wherever
we do a READ on it.