Applied subnet_id case-insensitive comparison to prevent PE recreation
Subnet ID resource IDs in Azure are case-insensitive, but the subnet_id comparison in the azurerm_private_endpoint is not.
If we call an azurerm_subnet data block and make the case mismatch the subnet_id of an azurerm_private_endpoint resource block where
the private endpoint already exists and the subnet_ids are functionally equivalent, the azurerm_private_endpoint will try to destry/create
the PE, which we do not want. Adding DiffSuppressFunc for case insensitivity fixes this.